requirements.txt
pytorch-lightning
openslide-python
shapely
opencv-python
albumentations